shmem_lock: Change usleep to equivalent load.
* It was noted that usleep(1) may lead to up to 50us delays.
To avoid that at the beginning of the test we calibrate the amount
of work equivalent to 1 us and use busyloop to emulate this delay.
This also better fits the nature of delay - access to the shared data.

NOTE: it was also noticed that mutexes don't obey FIFO policy when
acquiring locks. With higher rlock hold time this leads to really
writer starvation.